+++ src/Microsoft.VisualStudio.Composition/ExportDefinition.cs
@@ -9,6 +9,8 @@ using System.Diagnostics;
 using System.IO;
 using System.Reflection;
 using MessagePack;
+using MessagePack.Formatters;
+using Microsoft.VisualStudio.Composition.Formatter;
 
 [DebuggerDisplay("{" + nameof(ContractName) + ",nq}")]
 [MessagePackFormatter(typeof(ExportDefinitionFormatter))]
@@ -72,5 +74,57 @@ public class ExportDefinition : IEquatable<ExportDefinition>
         Requires.NotNull(assemblies, nameof(assemblies));
 
         ReflectionHelpers.GetInputAssembliesFromMetadata(assemblies, this.Metadata, nameGetter);
+    }
+
+    internal class ExportDefinitionFormatter : IMessagePackFormatter<ExportDefinition?>
+    {
+        public static readonly ExportDefinitionFormatter Instance = new();
+
+        private ExportDefinitionFormatter()
+        {
+        }
+
+        /// <inheritdoc/>
+        public ExportDefinition? Deserialize(ref MessagePackReader reader, MessagePackSerializerOptions options)
+        {
+            if (reader.TryReadNil())
+            {
+                return null;
+            }
+
+            options.Security.DepthStep(ref reader);
+
+            try
+            {
+                var actualCount = reader.ReadArrayHeader();
+                if (actualCount != 2)
+                {
+                    throw new MessagePackSerializationException($"Invalid array count for type {nameof(ExportDefinition)}. Expected: {2}, Actual: {actualCount}");
+                }
+
+                string contractName = options.Resolver.GetFormatterWithVerify<string>().Deserialize(ref reader, options);
+                IReadOnlyDictionary<string, object?> metadata = MetadataDictionaryFormatter.Instance.Deserialize(ref reader, options);
+
+                return new ExportDefinition(contractName, metadata);
+            }
+            finally
+            {
+                reader.Depth--;
+            }
+        }
+
+        public void Serialize(ref MessagePackWriter writer, ExportDefinition? value, MessagePackSerializerOptions options)
+        {
+            if (value is null)
+            {
+                writer.WriteNil();
+                return;
+            }
+
+            writer.WriteArrayHeader(2);
+
+            options.Resolver.GetFormatterWithVerify<string>().Serialize(ref writer, value.ContractName, options);
+            MetadataDictionaryFormatter.Instance.Serialize(ref writer, value.Metadata, options);
+        }
     }
 }
