+++ src/Microsoft.VisualStudio.Composition/ExportTypeIdentityConstraint.cs
@@ -11,6 +11,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using MessagePack;
+using MessagePack.Formatters;
 using Microsoft.VisualStudio.Composition.Formatter;
 using Microsoft.VisualStudio.Composition.Reflection;
 
@@ -73,5 +74,53 @@ public class ExportTypeIdentityConstraint : IImportSatisfiabilityConstraint, IDe
         }
 
         return this.TypeIdentityName == other.TypeIdentityName;
+    }
+
+    internal class ExportTypeIdentityConstraintFormatter : IMessagePackFormatter<ExportTypeIdentityConstraint?>
+    {
+        public static readonly ExportTypeIdentityConstraintFormatter Instance = new();
+
+        private ExportTypeIdentityConstraintFormatter()
+        {
+        }
+
+        public ExportTypeIdentityConstraint? Deserialize(ref MessagePackReader reader, MessagePackSerializerOptions options)
+        {
+            if (reader.TryReadNil())
+            {
+                return null;
+            }
+
+            options.Security.DepthStep(ref reader);
+
+            try
+            {
+                var actualCount = reader.ReadArrayHeader();
+                if (actualCount != 1)
+                {
+                    throw new MessagePackSerializationException($"Invalid array count for type {nameof(ExportTypeIdentityConstraint)}. Expected: {1}, Actual: {actualCount}");
+                }
+
+                string typeIdentityName = options.Resolver.GetFormatterWithVerify<string>().Deserialize(ref reader, options);
+                return new ExportTypeIdentityConstraint(typeIdentityName);
+            }
+            finally
+            {
+                reader.Depth--;
+            }
+        }
+
+        public void Serialize(ref MessagePackWriter writer, ExportTypeIdentityConstraint? value, MessagePackSerializerOptions options)
+        {
+            if (value is null)
+            {
+                writer.WriteNil();
+                return;
+            }
+
+            writer.WriteArrayHeader(1);
+
+            options.Resolver.GetFormatterWithVerify<string>().Serialize(ref writer, value.TypeIdentityName, options);
+        }
     }
 }
