+++ src/Microsoft.VisualStudio.Composition/ImportDefinition.cs
@@ -11,6 +11,7 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using MessagePack;
+using MessagePack.Formatters;
 using Microsoft.VisualStudio.Composition.Formatter;
 
 [DebuggerDisplay("{" + nameof(ContractName) + ",nq} ({Cardinality})")]
@@ -139,4 +140,65 @@ public class ImportDefinition : IEquatable<ImportDefinition>
         // TODO: consider the assembly dependencies brought in by constraints.
         ReflectionHelpers.GetInputAssembliesFromMetadata(assemblies, this.Metadata, nameRetriever);
     }
+
+    internal class ImportDefinitionFormatter : IMessagePackFormatter<ImportDefinition?>
+    {
+        public static readonly ImportDefinitionFormatter Instance = new();
+
+        private ImportDefinitionFormatter()
+        {
+        }
+
+        /// <inheritdoc/>
+        public ImportDefinition? Deserialize(ref MessagePackReader reader, MessagePackSerializerOptions options)
+        {
+            if (reader.TryReadNil())
+            {
+                return null;
+            }
+
+            options.Security.DepthStep(ref reader);
+
+            try
+            {
+                var actualCount = reader.ReadArrayHeader();
+                if (actualCount != 5)
+                {
+                    throw new MessagePackSerializationException($"Invalid array count for type {nameof(ImportDefinition)}. Expected: {5}, Actual: {actualCount}");
+                }
+
+                string contractName = options.Resolver.GetFormatterWithVerify<string>().Deserialize(ref reader, options);
+                ImportCardinality cardinality = options.Resolver.GetFormatterWithVerify<ImportCardinality>().Deserialize(ref reader, options);
+                IReadOnlyDictionary<string, object?> metadata = MetadataDictionaryFormatter.Instance.Deserialize(ref reader, options);
+
+                IReadOnlyList<IImportSatisfiabilityConstraint> constraints = options.Resolver.GetFormatterWithVerify<IReadOnlyList<IImportSatisfiabilityConstraint>>().Deserialize(ref reader, options);
+                IReadOnlyCollection<string> sharingBoundaries = options.Resolver.GetFormatterWithVerify<IReadOnlyCollection<string>>().Deserialize(ref reader, options);
+
+                return new ImportDefinition(contractName, cardinality, metadata, constraints, sharingBoundaries);
+            }
+            finally
+            {
+                reader.Depth--;
+            }
+        }
+
+        /// <inheritdoc/>
+        public void Serialize(ref MessagePackWriter writer, ImportDefinition? value, MessagePackSerializerOptions options)
+        {
+            if (value is null)
+            {
+                writer.WriteNil();
+                return;
+            }
+
+            writer.WriteArrayHeader(5);
+
+            options.Resolver.GetFormatterWithVerify<string>().Serialize(ref writer, value.ContractName, options);
+            options.Resolver.GetFormatterWithVerify<ImportCardinality>().Serialize(ref writer, value.Cardinality, options);
+            MetadataDictionaryFormatter.Instance.Serialize(ref writer, value.Metadata, options);
+            options.Resolver.GetFormatterWithVerify<IReadOnlyCollection<IImportSatisfiabilityConstraint>>().Serialize(ref writer, value.ExportConstraints, options);
+            options.Resolver.GetFormatterWithVerify<IReadOnlyCollection<string>>().Serialize(ref writer, value.ExportFactorySharingBoundaries, options);
+        }
+    }
+
 }
